Text Display after voice input progress
Text is being displayed to inform the user that their voice is being processed. However, the returned message is empty. This is probably because I have used the echoed message which appears to be empty. Will need to look into how the voice input is taken place in previous solution.