Added Documentation Systed
I have added a docs directory for the development branch, when adding a new feature to the project we should also be adding documentation where necessary to ensure the maintainability of the project.

This will provide a glanceable directory of information to refresh our memory when it comes to maintaining or building upon the project, or allow a different or new development team to jump right in to where they need to be without having to spend too long dissecting code to understand the logic.

The documentation needs to provide a high level overview of functionality as well as then providing in depth specifications for HTTP requests etc.

Using MKDocs, to use simply open the docs directory as a project in your ide(or in terminal/cmd) and run 'mkdocs serve' inside of the docs directory (not docs/docs) - this will run a local web server to render the documentation.

MKDocs allows you to write Documentation in Markdown and formats it into a website for you - saving time!