Updating AI selection screen
I have began tweaking the AI selection screen.

I have changed the layout to provide a selection box as well as an overview section of the scenario selected.

I am still working on the colour scheme and layout for the overview box.

A welcome message will be displayed for now while no scenario is selected.