Overhauled SelectAI screen
Entirely overhauled the SelectAI screen, now using MUI with a custom ThemeProvider to provide a sleek feel in-keeping with the general colour scheme we have been using so far.

Defined the them in file so we can easily change it if we need to.

We can, at a later date, extract the them into a wider wrapper component (such as in the router) should we need to, and use more custom themed MUI components throughout the wider project, if we need to but that was outside of the scope of this commit.