Added Homepage, Central Theme Manager, Updated Routing
In this commit I have changed the app routing slightly, added a central theme manager to ensure we can pass the MUI theme from the selectai page to all parts of the app.

As part of this I have updated the routing so the select ai has a more logical route, and so the default route corresponds to a homepage. I may make the select ai page a protected route, as well as the chat in the next update, so as to force login. However, this will be frontend validated - request tunelling could likely circumvent this, as could source manipulation. However, this will be fine for prototype purposes and backend intervention would come at a later stage.