Create useAuth.ts
useAuth hook - this can be modified to provide actual interaction with the backend to verify whether a user is authenticated, should this be something we decide to do for production.

The best approach in my opinion would be to have a wrapper around the fetch api, fetchWithAuth() for example - which is used for api requests on protected pages - this wrapper will attach the bearer authentication to the request automatically and can have in-built handling for if the user is not logged in. This allows restriction of data that is served from the backend when the user cannot be validated on both sides of the request.